requirements.txt
pytest==5.0.0
pycodestyle==2.5.0
behave==1.2.6
requests==2.24.0
slackclient==2.8.0